+++ src/main/java/com/misiac/workoutjournal/repository/EquipmentCategoryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface EquipmentCategoryRepository extends JpaRepository<EquipmentCategory, Integer> {
+public interface EquipmentCategoryRepository extends JpaRepository<EquipmentCategory, Long> {
 
     Optional<EquipmentCategory> findEquipmentCategoryByName(String name);
 }
