+++ src/main/java/com/misiac/workoutjournal/repository/MuscleGroupCategoryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MuscleGroupCategoryRepository extends JpaRepository<MuscleGroupCategory, Integer> {
+public interface MuscleGroupCategoryRepository extends JpaRepository<MuscleGroupCategory, Long> {
 
     Optional<MuscleGroupCategory> findMuscleGroupCategoryByName(String name);
 }
